Wire guest_config module in VMM
Includes moving cpuid module as a sub-module
for guest_config
