+++ src/vmm/src/guest_config/cpuid/indexing.rs
@@ -3,7 +3,8 @@
 #![allow(clippy::transmute_ptr_to_ptr, clippy::needless_lifetimes)]
 
 #[allow(clippy::wildcard_imports)]
-use crate::cpuid::leaves::*;
+use super::leaves::*;
+use crate::guest_config::cpuid::{AmdCpuid, CpuidEntry, IntelCpuid};
 
 /// Indexs leaf.
 pub trait IndexLeaf<const INDEX: usize> {
@@ -66,30 +67,30 @@ pub trait IndexLeafMut<const INDEX: usize> {
 /// ```
 macro_rules! index_leaf {
     ($index: literal, $leaf: ty, $cpuid: ty) => {
-        impl $crate::cpuid::IndexLeaf<$index> for $cpuid {
+        impl $crate::guest_config::cpuid::IndexLeaf<$index> for $cpuid {
             type Output<'a> = Option<&'a $leaf>;
             #[inline]
             fn index_leaf<'a>(&'a self) -> Self::Output<'a> {
                 self.0
-                    .get(&$crate::cpuid::CpuidKey::leaf($index))
+                    .get(&$crate::guest_config::cpuid::CpuidKey::leaf($index))
                     // JUSTIFICATION: There is no safe alternative.
                     // SAFETY: Transmuting references to same size and alignment types is safe. For
                     // further information See `index_leaf!`.
-                    .map(|entry: &'a crate::cpuid::CpuidEntry| unsafe {
+                    .map(|entry: &'a CpuidEntry| unsafe {
                         std::mem::transmute::<_, &'a $leaf>(&entry.result)
                     })
             }
         }
-        impl $crate::cpuid::IndexLeafMut<$index> for $cpuid {
+        impl $crate::guest_config::cpuid::IndexLeafMut<$index> for $cpuid {
             type Output<'a> = Option<&'a mut $leaf>;
             #[inline]
             fn index_leaf_mut<'a>(&'a mut self) -> Self::Output<'a> {
                 self.0
-                    .get_mut(&$crate::cpuid::CpuidKey::leaf($index))
+                    .get_mut(&$crate::guest_config::cpuid::CpuidKey::leaf($index))
                     // JUSTIFICATION: There is no safe alternative.
-                    // SAFETY: Transmuting references to same size and alignment types is safe. For
-                    // further information See `index_leaf!`.
-                    .map(|entry: &'a mut crate::cpuid::CpuidEntry| unsafe {
+                    // SAFETY: Transmuting references to same size and alignment types is safe.
+                    // For further information See `index_leaf!`.
+                    .map(|entry: &'a mut CpuidEntry| unsafe {
                         std::mem::transmute::<_, &'a mut $leaf>(&mut entry.result)
                     })
             }
@@ -102,36 +103,36 @@ pub(crate) use index_leaf;
 /// Convenience macro for indexing shared leaves.
 macro_rules! cpuid_index_leaf {
     ($index: literal, $leaf: ty) => {
-        impl IndexLeaf<$index> for super::Cpuid {
+        impl crate::guest_config::cpuid::IndexLeaf<$index> for super::Cpuid {
             type Output<'a> = Option<&'a $leaf>;
             #[inline]
             fn index_leaf<'a>(&'a self) -> Self::Output<'a> {
                 match self {
                     Self::Intel(intel_cpuid) => {
-                        <crate::cpuid::IntelCpuid as IndexLeaf<$index>>::index_leaf(intel_cpuid)
+                        <IntelCpuid as IndexLeaf<$index>>::index_leaf(intel_cpuid)
                     }
-                    Self::Amd(amd_cpuid) => {
-                        <crate::cpuid::AmdCpuid as IndexLeaf<$index>>::index_leaf(amd_cpuid)
-                    }
+                    Self::Amd(amd_cpuid) => <AmdCpuid as IndexLeaf<$index>>::index_leaf(amd_cpuid),
                 }
             }
         }
-        impl IndexLeafMut<$index> for crate::cpuid::Cpuid {
+        impl crate::guest_config::cpuid::IndexLeafMut<$index>
+            for crate::guest_config::cpuid::Cpuid
+        {
             type Output<'a> = Option<&'a mut $leaf>;
             #[inline]
             fn index_leaf_mut<'a>(&'a mut self) -> Self::Output<'a> {
                 match self {
-                    Self::Intel(intel_cpuid) => <crate::cpuid::IntelCpuid as IndexLeafMut<
-                        $index,
-                    >>::index_leaf_mut(intel_cpuid),
+                    Self::Intel(intel_cpuid) => {
+                        <IntelCpuid as IndexLeafMut<$index>>::index_leaf_mut(intel_cpuid)
+                    }
                     Self::Amd(amd_cpuid) => {
-                        <crate::cpuid::AmdCpuid as IndexLeafMut<$index>>::index_leaf_mut(amd_cpuid)
+                        <AmdCpuid as IndexLeafMut<$index>>::index_leaf_mut(amd_cpuid)
                     }
                 }
             }
         }
-        index_leaf!($index, $leaf, crate::cpuid::AmdCpuid);
-        index_leaf!($index, $leaf, crate::cpuid::IntelCpuid);
+        index_leaf!($index, $leaf, AmdCpuid);
+        index_leaf!($index, $leaf, IntelCpuid);
     };
 }
