+++ src/vmm/src/guest_config/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod cpuid;
+/// cpuid static template functions.
+pub(crate) mod static_templates;
