+++ src/vmm/src/guest_config/static_templates/mod.rs
