+++ src/vmm/src/guest_config/static_templates/t2s.rs
@@ -1,71 +1,78 @@
-// Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::cpuid::{AmdCpuid, Cpuid, CpuidEntry, CpuidKey, CpuidRegisters, KvmCpuidFlags};
+use kvm_bindings::kvm_msr_entry;
 
-/// This is translated from `cpuid -r` within a T2A guest microVM on an ec2 m6a.metal instance:
+use crate::arch::x86_64::msr::ArchCapaMSRFlags;
+use crate::arch_gen::x86::msr_index::MSR_IA32_ARCH_CAPABILITIES;
+use crate::guest_config::cpuid::cpuid_ffi::KvmCpuidFlags;
+use crate::guest_config::cpuid::{Cpuid, CpuidEntry, CpuidKey, CpuidRegisters, IntelCpuid};
+
+/// Add the MSR entries specific to this T2S template.
+#[inline]
+pub fn update_t2s_msr_entries(msr_entries: &mut Vec<kvm_msr_entry>) {
+    let capabilities = ArchCapaMSRFlags::RSBA
+        | ArchCapaMSRFlags::SKIP_L1DFL_VMENTRY
+        | ArchCapaMSRFlags::IF_PSCHANGE_MC_NO
+        | ArchCapaMSRFlags::MISC_PACKAGE_CTRLS
+        | ArchCapaMSRFlags::ENERGY_FILTERING_CTL;
+    msr_entries.push(kvm_msr_entry {
+        index: MSR_IA32_ARCH_CAPABILITIES,
+        data: capabilities.bits(),
+        ..kvm_msr_entry::default()
+    });
+}
+
+/// This is translated from `cpuid -r` within a T2S guest microVM on an ec2 m5.metal instance:
 ///
 /// ```text
 /// CPU 0:
-///    0x00000000 0x00: eax=0x00000010 ebx=0x68747541 ecx=0x444d4163 edx=0x69746e65
+///    0x00000000 0x00: eax=0x00000016 ebx=0x756e6547 ecx=0x6c65746e edx=0x49656e69
 ///    0x00000001 0x00: eax=0x000306f2 ebx=0x00020800 ecx=0xfffa3203 edx=0x178bfbff
-///    0x00000002 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000002 0x00: eax=0x76036301 ebx=0x00f0b5ff ecx=0x00000000 edx=0x00c30000
 ///    0x00000003 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000004 0x00: eax=0x04000121 ebx=0x01c0003f ecx=0x0000003f edx=0x00000000
+///    0x00000004 0x01: eax=0x04000122 ebx=0x01c0003f ecx=0x0000003f edx=0x00000000
+///    0x00000004 0x02: eax=0x04000143 ebx=0x03c0003f ecx=0x000003ff edx=0x00000000
+///    0x00000004 0x03: eax=0x04004163 ebx=0x0280003f ecx=0x0000bfff edx=0x00000004
 ///    0x00000005 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x00000006 0x00: eax=0x00000004 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x00000007 0x00: eax=0x00000000 ebx=0x001007ab ecx=0x00000000 edx=0x8c000000
+///    0x00000007 0x00: eax=0x00000000 ebx=0x001007ab ecx=0x00000000 edx=0xac000400
 ///    0x00000008 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x00000009 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x0000000a 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x0000000b 0x00: eax=0x00000000 ebx=0x00000001 ecx=0x00000100 edx=0x00000000
+///    0x0000000b 0x01: eax=0x00000007 ebx=0x00000002 ecx=0x00000201 edx=0x00000000
 ///    0x0000000c 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x0000000d 0x00: eax=0x00000007 ebx=0x00000340 ecx=0x00000988 edx=0x00000000
-///    0x0000000d 0x01: eax=0x00000001 ebx=0x00000348 ecx=0x00000000 edx=0x00000000
+///    0x0000000d 0x00: eax=0x00000007 ebx=0x00000340 ecx=0x00000a88 edx=0x00000000
+///    0x0000000d 0x01: eax=0x00000001 ebx=0x00000a08 ecx=0x00000000 edx=0x00000000
 ///    0x0000000d 0x02: eax=0x00000100 ebx=0x00000240 ecx=0x00000000 edx=0x00000000
 ///    0x0000000e 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x0000000f 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x00000010 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000011 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000012 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000013 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000014 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000015 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x00000016 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x40000000 0x00: eax=0x40000001 ebx=0x4b4d564b ecx=0x564b4d56 edx=0x0000004d
 ///    0x40000001 0x00: eax=0x01007efb ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000000 0x00: eax=0x8000001f ebx=0x68747541 ecx=0x444d4163 edx=0x69746e65
-///    0x80000001 0x00: eax=0x00a00f11 ebx=0x40000000 ecx=0x00c00237 edx=0x2813fbff
-///    0x80000002 0x00: eax=0x20444d41 ebx=0x43595045 ecx=0x00000000 edx=0x00000000
-///    0x80000003 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000004 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x80000000 0x00: eax=0x80000008 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
+///    0x80000001 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000021 edx=0x28100800
+///    0x80000002 0x00: eax=0x65746e49 ebx=0x2952286c ecx=0x6f655820 edx=0x2952286e
+///    0x80000003 0x00: eax=0x6f725020 ebx=0x73736563 ecx=0x4020726f edx=0x352e3220
+///    0x80000004 0x00: eax=0x7a484730 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0x80000005 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000006 0x00: eax=0x48002200 ebx=0x68004200 ecx=0x02006140 edx=0x06009140
+///    0x80000006 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x01006040 edx=0x00000000
 ///    0x80000007 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000100
-///    0x80000008 0x00: eax=0x00003030 ebx=0x030ed000 ecx=0x00007001 edx=0x00000000
-///    0x80000009 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000000a 0x00: eax=0x00000001 ebx=0x00000008 ecx=0x00000000 edx=0x00000009
-///    0x8000000b 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000000c 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000000d 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000000e 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000000f 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000010 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000011 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000012 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000013 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000014 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000015 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000016 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000017 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000018 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x80000019 0x00: eax=0xf040f040 ebx=0xf0400000 ecx=0x00000000 edx=0x00000000
-///    0x8000001a 0x00: eax=0x00000006 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000001b 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000001c 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000001d 0x00: eax=0x00000121 ebx=0x01c0003f ecx=0x0000003f edx=0x00000000
-///    0x8000001d 0x01: eax=0x00000122 ebx=0x01c0003f ecx=0x0000003f edx=0x00000000
-///    0x8000001d 0x02: eax=0x00000143 ebx=0x01c0003f ecx=0x000003ff edx=0x00000002
-///    0x8000001d 0x03: eax=0x00004163 ebx=0x03c0003f ecx=0x00007fff edx=0x00000001
-///    0x8000001e 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
-///    0x8000001f 0x00: eax=0x0101fd3f ebx=0x00004173 ecx=0x000001fd edx=0x000001fe
+///    0x80000008 0x00: eax=0x0000302e ebx=0x0100d000 ecx=0x00000000 edx=0x00000000
 ///    0x80860000 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 ///    0xc0000000 0x00: eax=0x00000000 ebx=0x00000000 ecx=0x00000000 edx=0x00000000
 /// ```
-pub fn template() -> Cpuid {
-    Cpuid::Amd(AmdCpuid({
+#[allow(clippy::too_many_lines)]
+pub fn t2s() -> Cpuid {
+    Cpuid::Intel(IntelCpuid({
         let mut map = std::collections::BTreeMap::new();
         map.insert(
             CpuidKey {
@@ -75,10 +82,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x10,
-                    ebx: 0x68747541,
-                    ecx: 0x444d4163,
-                    edx: 0x69746e65,
+                    eax: 0x16,
+                    ebx: 0x756e6547,
+                    ecx: 0x6c65746e,
+                    edx: 0x49656e69,
                 },
             },
         );
@@ -92,7 +99,7 @@ pub fn template() -> Cpuid {
                 result: CpuidRegisters {
                     eax: 0x306f2,
                     ebx: 0x20800,
-                    ecx: 0xfffa3203,
+                    ecx: 0xf7fa3203,
                     edx: 0x178bfbff,
                 },
             },
@@ -105,10 +112,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
+                    eax: 0x76036301,
+                    ebx: 0xf0b5ff,
                     ecx: 0x0,
-                    edx: 0x0,
+                    edx: 0xc30000,
                 },
             },
         );
@@ -124,6 +131,66 @@ pub fn template() -> Cpuid {
                     ebx: 0x0,
                     ecx: 0x0,
                     edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x4,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x4000121,
+                    ebx: 0x1c0003f,
+                    ecx: 0x3f,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x4,
+                subleaf: 0x1,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x4000122,
+                    ebx: 0x1c0003f,
+                    ecx: 0x3f,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x4,
+                subleaf: 0x2,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x4000143,
+                    ebx: 0x3c0003f,
+                    ecx: 0x3ff,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x4,
+                subleaf: 0x3,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x4004163,
+                    ebx: 0x280003f,
+                    ecx: 0xbfff,
+                    edx: 0x4,
                 },
             },
         );
@@ -168,7 +235,7 @@ pub fn template() -> Cpuid {
                     eax: 0x0,
                     ebx: 0x1007ab,
                     ecx: 0x0,
-                    edx: 0x8c000000,
+                    edx: 0xac000400,
                 },
             },
         );
@@ -219,6 +286,36 @@ pub fn template() -> Cpuid {
         );
         map.insert(
             CpuidKey {
+                leaf: 0xb,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x1,
+                    ecx: 0x100,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0xb,
+                subleaf: 0x1,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x7,
+                    ebx: 0x2,
+                    ecx: 0x201,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
                 leaf: 0xc,
                 subleaf: 0x0,
             },
@@ -242,7 +339,7 @@ pub fn template() -> Cpuid {
                 result: CpuidRegisters {
                     eax: 0x7,
                     ebx: 0x340,
-                    ecx: 0x988,
+                    ecx: 0xa88,
                     edx: 0x0,
                 },
             },
@@ -256,7 +353,7 @@ pub fn template() -> Cpuid {
                 flags: KvmCpuidFlags(0x1),
                 result: CpuidRegisters {
                     eax: 0x1,
-                    ebx: 0x348,
+                    ebx: 0xa08,
                     ecx: 0x0,
                     edx: 0x0,
                 },
@@ -298,7 +395,7 @@ pub fn template() -> Cpuid {
                 subleaf: 0x0,
             },
             CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
+                flags: KvmCpuidFlags(0x1),
                 result: CpuidRegisters {
                     eax: 0x0,
                     ebx: 0x0,
@@ -310,6 +407,96 @@ pub fn template() -> Cpuid {
         map.insert(
             CpuidKey {
                 leaf: 0x10,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x11,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x0),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x12,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x13,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x0),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x14,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x1),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x15,
+                subleaf: 0x0,
+            },
+            CpuidEntry {
+                flags: KvmCpuidFlags(0x0),
+                result: CpuidRegisters {
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
+                },
+            },
+        );
+        map.insert(
+            CpuidKey {
+                leaf: 0x16,
                 subleaf: 0x0,
             },
             CpuidEntry {
@@ -345,7 +532,7 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x01007efb,
+                    eax: 0x1007efb,
                     ebx: 0x0,
                     ecx: 0x0,
                     edx: 0x0,
@@ -360,10 +547,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x8000001f,
-                    ebx: 0x68747541,
-                    ecx: 0x444d4163,
-                    edx: 0x69746e65,
+                    eax: 0x80000008,
+                    ebx: 0x0,
+                    ecx: 0x0,
+                    edx: 0x0,
                 },
             },
         );
@@ -375,10 +562,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0xa00f11,
-                    ebx: 0x40000000,
-                    ecx: 0xc00237,
-                    edx: 0x2813fbff,
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x21,
+                    edx: 0x28100800,
                 },
             },
         );
@@ -390,10 +577,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x20444d41,
-                    ebx: 0x43595045,
-                    ecx: 0x0,
-                    edx: 0x0,
+                    eax: 0x65746e49,
+                    ebx: 0x2952286c,
+                    ecx: 0x6f655820,
+                    edx: 0x2952286e,
                 },
             },
         );
@@ -405,10 +592,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
+                    eax: 0x6f725020,
+                    ebx: 0x73736563,
+                    ecx: 0x4020726f,
+                    edx: 0x302e3320,
                 },
             },
         );
@@ -420,7 +607,7 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x0,
+                    eax: 0x7a484730,
                     ebx: 0x0,
                     ecx: 0x0,
                     edx: 0x0,
@@ -450,10 +637,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x48002200,
-                    ebx: 0x68004200,
-                    ecx: 0x2006140,
-                    edx: 0x6009140,
+                    eax: 0x0,
+                    ebx: 0x0,
+                    ecx: 0x1006040,
+                    edx: 0x0,
                 },
             },
         );
@@ -480,415 +667,10 @@ pub fn template() -> Cpuid {
             CpuidEntry {
                 flags: KvmCpuidFlags(0x0),
                 result: CpuidRegisters {
-                    eax: 0x3030,
-                    ebx: 0x30ed000,
-                    ecx: 0x7001,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000009,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
+                    eax: 0x302e,
+                    ebx: 0x100d000,
                     ecx: 0x0,
                     edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0xc000000a,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x1,
-                    ebx: 0x8,
-                    ecx: 0x0,
-                    edx: 0x9,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000000b,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000000c,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000000d,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000000e,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000000f,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000010,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000011,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000012,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000013,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000014,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000015,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000016,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000017,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000018,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x80000019,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0xf040f040,
-                    ebx: 0xf0400000,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001a,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x6,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001b,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x6,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001c,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x6,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001d,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x121,
-                    ebx: 0x1c0003f,
-                    ecx: 0x3f,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001d,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x1),
-                result: CpuidRegisters {
-                    eax: 0x121,
-                    ebx: 0x1c0003f,
-                    ecx: 0x3f,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001d,
-                subleaf: 0x1,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x1),
-                result: CpuidRegisters {
-                    eax: 0x122,
-                    ebx: 0x1c0003f,
-                    ecx: 0x3f,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001d,
-                subleaf: 0x2,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x1),
-                result: CpuidRegisters {
-                    eax: 0x143,
-                    ebx: 0x1c0003f,
-                    ecx: 0x3ff,
-                    edx: 0x2,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001d,
-                subleaf: 0x3,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x1),
-                result: CpuidRegisters {
-                    eax: 0x4163,
-                    ebx: 0x3c0003f,
-                    ecx: 0x7fff,
-                    edx: 0x1,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001e,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x0,
-                    ebx: 0x0,
-                    ecx: 0x0,
-                    edx: 0x0,
-                },
-            },
-        );
-        map.insert(
-            CpuidKey {
-                leaf: 0x8000001f,
-                subleaf: 0x0,
-            },
-            CpuidEntry {
-                flags: KvmCpuidFlags(0x0),
-                result: CpuidRegisters {
-                    eax: 0x101fd3f,
-                    ebx: 0x4173,
-                    ecx: 0x1fd,
-                    edx: 0x1fe,
                 },
             },
         );
