+++ src/vmm/src/persist.rs
@@ -25,11 +25,11 @@ use vm_memory::{GuestMemory, GuestMemoryMmap};
 #[cfg(target_arch = "aarch64")]
 use crate::arch::regs::{get_manufacturer_id_from_host, get_manufacturer_id_from_state};
 use crate::builder::{self, BuildMicrovmFromSnapshotError};
-#[cfg(target_arch = "x86_64")]
-use crate::cpuid::common::get_vendor_id_from_host;
-#[cfg(target_arch = "x86_64")]
-use crate::cpuid::CpuidTrait;
 use crate::device_manager::persist::{DeviceStates, Error as DevicePersistError};
+#[cfg(target_arch = "x86_64")]
+use crate::guest_config::cpuid::common::get_vendor_id_from_host;
+#[cfg(target_arch = "x86_64")]
+use crate::guest_config::cpuid::CpuidTrait;
 use crate::memory_snapshot::{GuestMemoryState, SnapshotMemory};
 use crate::resources::VmResources;
 #[cfg(target_arch = "x86_64")]
@@ -349,10 +349,10 @@ pub fn get_snapshot_data_version(
 pub enum ValidateCpuVendorError {
     /// Failed to read host vendor.
     #[error("Failed to read host vendor: {0}")]
-    Host(#[from] crate::cpuid::common::GetCpuidError),
+    Host(#[from] crate::guest_config::cpuid::common::GetCpuidError),
     /// Failed to read snapshot vendor.
     #[error("Failed to read snapshot vendor: {0}")]
-    Snapshot(#[from] crate::cpuid::common::Leaf0NotFoundInCpuid),
+    Snapshot(#[from] crate::guest_config::cpuid::common::Leaf0NotFoundInCpuid),
 }
 
 /// Validates that snapshot CPU vendor matches the host CPU vendor.
@@ -371,7 +371,7 @@ pub fn validate_cpu_vendor(
     let snapshot_vendor_id = microvm_state.vcpu_states[0]
         .cpuid
         .vendor_id()
-        .ok_or(crate::cpuid::common::Leaf0NotFoundInCpuid)
+        .ok_or(crate::guest_config::cpuid::common::Leaf0NotFoundInCpuid)
         .map_err(|err| {
             error!("Snapshot CPU vendor is missing.");
             err
