--- src/vmm/src/vstate/vcpu/x86_64/cpuid_templates/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod c3;
-pub mod t2;
-pub mod t2a;
-pub mod t2cl;
-pub mod t2s;
