+++ src/vmm/tests/integration_tests.rs
@@ -379,7 +379,7 @@ fn test_snapshot_cpu_vendor_mismatch() {
     assert_eq!(
         validate_cpu_vendor(&microvm_state),
         Err(vmm::persist::ValidateCpuVendorError::Snapshot(
-            vmm::cpuid::common::Leaf0NotFoundInCpuid
+            vmm::guest_config::cpuid::common::Leaf0NotFoundInCpuid
         ))
     );
 }
